conslaws-poc_v1/requirements.txt
opensearch-py>=2.5.0
faiss-cpu>=1.8.0
sentence-transformers>=2.2.2
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.6.0
python-dotenv>=1.0.1
numpy>=1.26
pandas>=2.2
scikit-learn>=1.3
httpx>=0.27
openai>=1.40.0
streamlit>=1.37.0
PyYAML>=6.0
datasets>=2.14.0
ragas>=0.1.6
evaluate>=0.4.2
tqdm>=4.66.0
